- EMAR-227: Patient API updates for additional fields: image, numbers Emar.Api\Helpers\MediaTypes.cs Emar.Core\Helpers\Utilities.cs Emar.Core\Patients\Model\Constants.cs Emar.Core\Patients\Model\PatientDto.cs Emar.Core\Patients\Model\Mappings\PatientMapper.cs Emar.Core\Patients\Repository\IPatientRepository.cs Emar.Core\Patients\Repository\PatientRepository.cs Emar.Core\Patients\Service\IPatientService.cs Emar.Core\Patients\Service\PatientService.cs Emar.Data\Entities\ExternalId.cs Emar.Data\Entities\Patient.cs Emar.Data\Entities\Site.cs
- No ticket: Switch SiteId from int to long
Emar.Core\Medications\Model\MedicationRouteDto.cs
Emar.Core\Medications\Model\MedicationUnitDto.cs
Emar.Core\Options\Model\OptionDto.cs
Emar.Core\Options\Repository\IOptionRepository.cs
Emar.Core\Options\Repository\OptionRepository.cs
Emar.Core\Options\Service\IOptionService.cs
Emar.Core\Options\Service\OptionService.cs
Emar.Core\Orders\Model\UserQuickListItemDto.cs
Emar.Core\Orders\Repository\IOrderRepository.cs
Emar.Core\Orders\Repository\OrderRepository.cs
Emar.Core\Orders\Service\IOrderService.cs
Emar.Core\Orders\Service\OrderService.cs
Emar.Core\Orders\Service\OrderService_MockMethods.cs
Emar.Core\Patients\Model\PatientDto.cs
Emar.Core\ResourceParameters\PatientsResourceParameters.cs
Emar.Core\Sites\Model\SiteDto.cs
Emar.Data\Entities\Action.cs
Emar.Data\Entities\GroupListItem.cs
Emar.Data\Entities\MedicationRoute.cs
Emar.Data\Entities\MedicationUnit.cs
Emar.Data\Entities\Patient.cs
Emar.Data\Entities\Site.cs
Emar.Data\Entities\SiteOption.cs
Emar.Data\Entities\User.cs
Emar.Data\Entities\UserQuickListItem.cs

- No ticket: Cleanup
Emar.Data\EmarContext.cs

git-tfs-id: [http://ros-tfs-px01.picis.com:8080/tfs/emar]$/;C330

+++ WebAPI/API/C_Development/Work/Antoni/Emar.Core/Orders/Repository/IOrderRepository.cs
@@ -14,8 +14,8 @@ namespace Emar.Core.Orders.Repository
         IEnumerable<OrderEvent> GetEvents(long orderId);
         OrderEvent GetEvent(long eventId);
         IEnumerable<OrderEvent> GetAdministrationEvents(long administrationId);
-        IEnumerable<UserQuickListItem> GetUserQuickListMostUsed(int userId, int? siteId);
-        List<string> GetUserQuickListTabs(int userId, int? siteId);
-        IEnumerable<UserQuickListItem> GetUserQuickListTabItems(int userId, int? siteId, string tab);
+        IEnumerable<UserQuickListItem> GetUserQuickListMostUsed(int userId, long? siteId);
+        List<string> GetUserQuickListTabs(int userId, long? siteId);
+        IEnumerable<UserQuickListItem> GetUserQuickListTabItems(int userId, long? siteId, string tab);
     }
 }
